requirements.txt
Django==5.0
django-htmx
python-dotenv
psycopg2-binary
django-ulid
python-ulid
django-template-partials
